Moved geometry classes to Utility/Geometry
